+++ packages/core/datetime-picker/src/internal/index.js
@@ -4,6 +4,10 @@ import React from 'react';
 
 export const ClearIndicator = null;
 
+// This date was chosen to clearly show date and time formats (day > 12)
+// e.g. 18/02/1993 vs. 2/18/1993 and 1:00 PM vs 13:00
+export const placeholderDatetime = new Date(1993, 1, 18, 13);
+
 export const defaultTimes = [
   '09:00',
   '09:30',