+++ packages/help/help/src/components/Help.tsx
@@ -15,6 +15,8 @@ import HelpContent from './HelpContent';
 export interface Props extends WithAnalyticsEventsProps {
   // Id of the article to display. This prop is optional, if is not defined the default content will be displayed
   articleId?: string;
+  // Setter for the articleId. This prop is optional, if is not defined, the back button will not be visible
+  articleIdSetter?(id: string): void;
   // Function used to get an article content. This prop is optional, if is not defined the default content will be displayed
   onGetArticle?(id: string): Promise<Article>;
   // Function used to search an article.  This prop is optional, if is not defined search input will be hidden
@@ -39,16 +41,18 @@ export interface Props extends WithAnalyticsEventsProps {
     event?: React.MouseEvent<HTMLElement, MouseEvent>,
     analyticsEvent?: UIAnalyticsEvent,
   ): void;
+  // Footer content. This prop is optional
+  footer?: React.ReactNode;
   // Default content. This prop is optional
   children?: React.ReactNode;
 }
 
 export class Help extends React.Component<Props> {
   render() {
-    const { children, ...rest } = this.props;
+    const { children, footer, ...rest } = this.props;
 
     return (
-      <HelpContextProvider {...rest} defaultContent={children}>
+      <HelpContextProvider {...rest} defaultContent={children} footer={footer}>
         <MessagesIntlProvider>
           <HelpContent />
         </MessagesIntlProvider>