+++ packages/navigation/app-navigation/src/common/styles.ts
@@ -1,20 +1,19 @@
-import { css } from '@emotion/core';
 import { AppNavigationTheme } from '../theme';
 import { CREATE_BREAKPOINT } from './constants';
 
-export const actionSectionDesktopStyles = css`
-  @media (max-width: ${CREATE_BREAKPOINT - 1}px) {
-    display: none;
-  }
-`;
+export const actionSectionDesktopCSS = {
+  [`@media (max-width: ${CREATE_BREAKPOINT - 1}px)`]: {
+    display: 'none',
+  },
+};
 
-export const actionSectionMobileStyles = css`
-  @media (min-width: ${CREATE_BREAKPOINT}px) {
-    display: none;
-  }
-`;
+export const actionSectionMobileCSS = {
+  [`@media (min-width: ${CREATE_BREAKPOINT}px)`]: {
+    display: 'none',
+  },
+};
 
-export const skeletonStyles = ({ mode: { skeleton } }: AppNavigationTheme) => ({
+export const skeletonCSS = ({ mode: { skeleton } }: AppNavigationTheme) => ({
   opacity: 0.15,
   ...skeleton,
 });