+++ packages/navigation/app-navigation/src/components/BadgeContainer/index.tsx
@@ -1,15 +1,14 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import { styles } from './styles';
+import { badgeContainerCSS, containerCSS } from './styles';
 import { BadgeProps } from './types';
 
 export const BadgeContainer = (props: BadgeProps) => {
   const { badge: Badge, children } = props;
   return (
-    // Set z-index: 1 to let the badge (if any) overlap other items.
-    <div css={styles.container}>
+    <div css={containerCSS}>
       {children}
-      <div css={styles.badgePositioner}>
+      <div css={badgeContainerCSS}>
         <Badge />
       </div>
     </div>