+++ packages/navigation/app-navigation/src/components/Create/skeleton.tsx
@@ -2,15 +2,20 @@
 import { gridSize } from '@atlaskit/theme/constants';
 import { jsx } from '@emotion/core';
 import { Fragment } from 'react';
-import { IconButtonSkeleton } from '../IconButton/skeleton';
-import { createButtonSkeletonStyles, createIconSkeletonStyles } from './styles';
 
-export const CreateSkeleton = () => (
-  <Fragment>
-    <div css={createButtonSkeletonStyles} />
-    <IconButtonSkeleton
-      css={createIconSkeletonStyles}
-      dimension={gridSize() * 3.25}
-    />
-  </Fragment>
-);
+import { useTheme } from '../../theme';
+import { IconButtonSkeleton } from '../IconButton/skeleton';
+import { createButtonSkeletonCSS, createIconSkeletonCSS } from './styles';
+
+export const CreateSkeleton = () => {
+  const theme = useTheme();
+  return (
+    <Fragment>
+      <div css={createButtonSkeletonCSS(theme)} />
+      <IconButtonSkeleton
+        css={createIconSkeletonCSS}
+        size={gridSize() * 3.25}
+      />
+    </Fragment>
+  );
+};