+++ packages/navigation/app-navigation/src/components/IconButton/index.tsx
@@ -3,7 +3,7 @@ import { jsx } from '@emotion/core';
 import Button from '@atlaskit/button';
 import Tooltip from '@atlaskit/tooltip';
 
-import { withAppNavigationTheme } from '../../theme';
+import { useTheme } from '../../theme';
 import { getIconButtonTheme } from './styles';
 import { IconButtonProps } from './types';
 
@@ -12,19 +12,20 @@ export const IconButton = ({
   icon,
   onClick,
   testId = 'NavigationItem',
-  theme,
   tooltip,
-}: IconButtonProps) => (
-  <Tooltip content={tooltip}>
-    <Button
-      appearance="primary"
-      data-test-id={testId}
-      className={className}
-      iconBefore={icon}
-      onClick={onClick}
-      theme={getIconButtonTheme(theme)}
-    />
-  </Tooltip>
-);
+}: IconButtonProps) => {
+  const theme = useTheme();
 
-export const ThemedIconButton = withAppNavigationTheme(IconButton);
+  return (
+    <Tooltip content={tooltip}>
+      <Button
+        appearance="primary"
+        data-test-id={testId}
+        className={className}
+        iconBefore={icon}
+        onClick={onClick}
+        theme={getIconButtonTheme(theme)}
+      />
+    </Tooltip>
+  );
+};