+++ packages/navigation/app-navigation/src/components/PrimaryButton/styles.ts
@@ -3,15 +3,15 @@ import {
   fontSizeSmall,
   gridSize as gridSizeFn,
 } from '@atlaskit/theme/constants';
-import css from '@emotion/css';
+import { skeletonCSS } from '../../common/styles';
 import { AppNavigationTheme } from '../../theme';
 
 const gridSize = gridSizeFn();
 
-export const chevronStyles = css`
-  margin: 0 -${gridSize}px;
-  visibility: hidden;
-`;
+export const chevronCSS = {
+  margin: `0 -${gridSize}px`,
+  visibility: 'hidden' as const,
+};
 
 export const buttonHeight = gridSize * 4;
 
@@ -34,11 +34,9 @@ export const getPrimaryButtonTheme = ({
   return {
     buttonStyles: {
       ...buttonStyles,
-      display: 'inline-flex',
       fontSize: fontSizeSmall(),
       fontWeight: 'bold',
       height: buttonHeight,
-      marginLeft: margin.left,
       padding: padding.all,
       textTransform: 'uppercase',
       ...primaryButton.default,
@@ -52,3 +50,11 @@ export const getPrimaryButtonTheme = ({
     spinnerStyles,
   };
 };
+
+export const primaryButtonSkeletonCSS = (theme: AppNavigationTheme) => ({
+  borderRadius: `${gridSize / 2}px`,
+  display: 'inline-flex',
+  height: `${buttonHeight - padding.all * 2.5}px`,
+  width: '68px',
+  ...skeletonCSS(theme),
+});