+++ packages/navigation/app-navigation/src/components/ProductHome/skeleton.tsx
@@ -1,14 +1,19 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
+import { useTheme } from '../../theme';
 import {
-  containerSkeletonStyles,
-  productIconSkeletonStyles,
-  productLogoSkeletonStyles,
+  containerSkeletonCSS,
+  productIconSkeletonCSS,
+  productLogoSkeletonCSS,
 } from './styles';
 
-export const ProductHomeSkeleton = () => (
-  <div css={containerSkeletonStyles}>
-    <div css={productLogoSkeletonStyles} />
-    <div css={productIconSkeletonStyles} />
-  </div>
-);
+export const ProductHomeSkeleton = () => {
+  const theme = useTheme();
+
+  return (
+    <div css={containerSkeletonCSS}>
+      <div css={productLogoSkeletonCSS(theme)} />
+      <div css={productIconSkeletonCSS(theme)} />
+    </div>
+  );
+};