+++ packages/navigation/app-navigation/src/components/Search/skeleton.tsx
@@ -3,23 +3,26 @@ import { gridSize } from '@atlaskit/theme/constants';
 import { jsx } from '@emotion/core';
 import { Fragment } from 'react';
 
+import { useTheme } from '../../theme';
 import { IconButtonSkeleton } from '../IconButton/skeleton';
 import {
-  searchIconSkeletonStyles,
-  searchInputContainerStyles,
-  searchInputSkeletonStyles,
+  searchIconSkeletonCSS,
+  searchInputContainerCSS,
+  searchInputSkeletonCSS,
 } from './styles';
 
 export const SearchSkeleton = () => {
+  const theme = useTheme();
+
   return (
     <Fragment>
-      <div css={searchInputContainerStyles}>
-        <div css={searchInputSkeletonStyles} />
+      <div css={searchInputContainerCSS}>
+        <div css={searchInputSkeletonCSS(theme)} />
       </div>
       <IconButtonSkeleton
-        css={searchIconSkeletonStyles}
-        dimension={gridSize() * 3.25}
+        css={searchIconSkeletonCSS}
         marginRight={5}
+        size={gridSize() * 3.25}
       />
     </Fragment>
   );