+++ packages/navigation/atlassian-switcher/src/__tests__/unit/map-results-to-switcher-props.ts
@@ -28,6 +28,7 @@ describe('map-results-to-switcher-props', () => {
           productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: true,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({ sites: [] }),
       );
@@ -54,6 +55,7 @@ describe('map-results-to-switcher-props', () => {
           xflow: true,
           productTopItemVariation: ProductTopItemVariation.currentSite,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({ sites: [] }),
       );
@@ -80,6 +82,7 @@ describe('map-results-to-switcher-props', () => {
           productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: true,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({ sites: [] }),
       );
@@ -110,6 +113,7 @@ describe('map-results-to-switcher-props', () => {
           xflow: true,
           productTopItemVariation: ProductTopItemVariation.currentSite,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({ sites: [] }),
       );
@@ -136,6 +140,7 @@ describe('map-results-to-switcher-props', () => {
           productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: true,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({ sites: [] }),
       );
@@ -160,6 +165,7 @@ describe('map-results-to-switcher-props', () => {
           productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: false,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({ sites: [] }),
       );
@@ -182,6 +188,7 @@ describe('map-results-to-switcher-props', () => {
           xflow: false,
           isDiscoverMoreForEveryoneEnabled: false,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({
           sites: [
@@ -261,6 +268,7 @@ describe('map-results-to-switcher-props', () => {
           xflow: false,
           isDiscoverMoreForEveryoneEnabled: false,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({
           sites: [
@@ -315,6 +323,7 @@ describe('map-results-to-switcher-props', () => {
           xflow: false,
           isDiscoverMoreForEveryoneEnabled: false,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({
           sites: [
@@ -369,6 +378,7 @@ describe('map-results-to-switcher-props', () => {
           xflow: false,
           isDiscoverMoreForEveryoneEnabled: false,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({
           sites: [
@@ -402,6 +412,7 @@ describe('map-results-to-switcher-props', () => {
           productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: false,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({
           sites: [
@@ -433,6 +444,7 @@ describe('map-results-to-switcher-props', () => {
           productTopItemVariation: ProductTopItemVariation.currentSite,
           isDiscoverMoreForEveryoneEnabled: false,
           disableHeadings: false,
+          isEmceeLinkEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({
           sites: [