+++ packages/navigation/atlassian-switcher/src/index.ts
@@ -1,5 +1,5 @@
 export { SwitchToTooltipText } from './utils/messages';
-export { AtlassianSwitcherLoader as default } from './loaders';
+export { AtlassianSwitcherLoader as default } from './components/loaders';
 export {
   default as AtlassianSwitcherPrefetchTrigger,
 } from './components/prefetch-trigger';