+++ packages/media/media-picker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/media-picker
 
+## 47.1.3
+
+### Patch Changes
+
+- [patch][fc79969f86](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/fc79969f86):
+
+  Update all the theme imports in media to use multi entry points
+
 ## 47.1.2
 
 - Updated dependencies [af72468517](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/af72468517):